+++ .ENV
@@ -1,6 +1,10 @@
-
 # Supabase Project URL
 VITE_SUPABASE_URL=https://zzqoxgytfrbptojcwrjm.supabase.co
+SUPABASE_URL=https://zzqoxgytfrbptojcwrjm.supabase.co
 
 # Supabase Public Anon Key
 VITE_SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Inp6cW94Z3l0ZnJicHRvamN3cmptIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDcxMzQ5OTYsImV4cCI6MjA2MjcxMDk5Nn0.Tv_3Na5FFM35FI-p7vIlNj3yAo_WqkKAqLn3fqQ6k3Q
+SUPABASE_ANON_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Inp6cW94Z3l0ZnJicHRvamN3cmptIiwicm9sZSI6ImFub24iLCJpYXQiOjE3NDcxMzQ5OTYsImV4cCI6MjA2MjcxMDk5Nn0.Tv_3Na5FFM35FI-p7vIlNj3yAo_WqkKAqLn3fqQ6k3Q
+
+# Supabase Service Role Key
+SUPABASE_SERVICE_ROLE_KEY=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJzdXBhYmFzZSIsInJlZiI6Inp6cW94Z3l0ZnJicHRvamN3cmptIiwicm9sZSI6InNlcnZpY2Vfcm9sZSIsImlhdCI6MTc0NzEzNDk5NiwiZXhwIjoyMDYyNzEwOTk2fQ.nR8TR_Kyk5SucU-KOpWIQ2zu-UasYSJTS8OhJGXjbGE
