+++ backend/routers/alliance_changelog.py
@@ -4,17 +4,7 @@ from datetime import datetime
 from .progression_router import get_user_id
 
 
-def get_supabase_client():
-    try:
-        from supabase import create_client
-    except ImportError as e:
-        raise RuntimeError("supabase client library not installed") from e
-    import os
-    url = os.getenv("SUPABASE_URL")
-    key = os.getenv("SUPABASE_SERVICE_ROLE_KEY") or os.getenv("SUPABASE_KEY")
-    if not url or not key:
-        raise RuntimeError("Supabase credentials not configured")
-    return create_client(url, key)
+from ..supabase_client import get_supabase_client
 
 
 router = APIRouter(prefix="/api/alliance/changelog", tags=["alliance_changelog"])
