+++ backend/routers/alliance_members_view.py
@@ -9,17 +9,7 @@ def get_current_user_id(x_user_id: str | None = Header(None)) -> str:
     return x_user_id
 
 
-def get_supabase_client():
-    try:
-        from supabase import create_client
-    except ImportError as e:
-        raise RuntimeError("supabase client library not installed") from e
-    import os
-    url = os.getenv("SUPABASE_URL")
-    key = os.getenv("SUPABASE_SERVICE_ROLE_KEY") or os.getenv("SUPABASE_KEY")
-    if not url or not key:
-        raise RuntimeError("Supabase credentials not configured")
-    return create_client(url, key)
+from ..supabase_client import get_supabase_client
 
 
 @router.get("/api/alliance-members/view")
