+++ backend/routers/treaty_web.py
@@ -2,17 +2,7 @@ from fastapi import APIRouter, Depends
 from ..security import verify_jwt_token
 
 
-def get_supabase_client():
-    try:
-        from supabase import create_client
-    except ImportError as e:  # pragma: no cover
-        raise RuntimeError("supabase client library not installed") from e
-    import os
-    url = os.getenv("SUPABASE_URL")
-    key = os.getenv("SUPABASE_SERVICE_ROLE_KEY") or os.getenv("SUPABASE_KEY")
-    if not url or not key:
-        raise RuntimeError("Supabase credentials not configured")
-    return create_client(url, key)
+from ..supabase_client import get_supabase_client
 
 
 router = APIRouter(prefix="/api/treaty_web", tags=["treaty_web"])
